handled errors

+++ src/main/java/lab/microservice/controller/Controller.java
@@ -32,18 +32,25 @@ public class Controller {
 
     @PostMapping
     public ResponseEntity<Item> createItem(@RequestBody Item item) {
+        if (service.getAllItems().stream().map(Item::getName).filter(name -> item.getName().equals(name)).toList().size() > 0) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         Item createdItem = service.createItem(item);
         return new ResponseEntity<>(createdItem, HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<Item> updateItem(@PathVariable Long id, @RequestBody Item itemDetails) {
+        if (service.getItemById(id).isEmpty()) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         Item updatedItem = service.updateItem(id, itemDetails);
         return new ResponseEntity<>(updatedItem, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteItem(@PathVariable Long id) {
+
+        if (service.getItemById(id).isEmpty()) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
         service.deleteItem(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
